# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AZOX Chess API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  waiting: 'waiting',
  active: 'active',
  finished: 'finished',
} as const;

export interface Match {
  /** Unique 6-character match code */
  code: string;
  status: MatchStatus;
  /** Number of players connected (0, 1, or 2) */
  playerCount?: number;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}