# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AZOX Chess API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: matches
    description: Chess match management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /matches:
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a new multiplayer match
      description: Generates a unique match code for the host
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"

  /matches/{code}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match status
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Match not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /matches/{code}/join:
    post:
      operationId: joinMatch
      tags: [matches]
      summary: Join an existing match using a code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Match not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Match already full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Match:
      type: object
      required:
        - code
        - status
        - createdAt
      properties:
        code:
          type: string
          description: Unique 6-character match code
        status:
          type: string
          enum: [waiting, active, finished]
        playerCount:
          type: integer
          description: Number of players connected (0, 1, or 2)
        createdAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string