# lib/api-zod/src/generated/types/match.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AZOX Chess API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStatus } from './matchStatus';

export interface Match {
  /** Unique 6-character match code */
  code: string;
  status: MatchStatus;
  /** Number of players connected (0, 1, or 2) */
  playerCount?: number;
  createdAt: Date;
}